==> 04-missing.Rmd <==
# Missing values

## Missing values by year
We first examine the missing condition in each year. We draw a plot to show the total number of missing values in each year
```{r ch04-1, fig.height=4}
tsize <- 15

df_countryData_selected <- read_csv("./WDI_csv/WDI_countryData_selected.csv")
df_series <- read_csv("./WDI_csv/WDISeries.csv")

missing_by_year <- colSums(is.na(df_countryData_selected %>% 
                select(-c("Country Name", "Country Code", "Indicator Name", "Indicator Code")))) 
df_missing_by_year <- data.frame(year=1960:2020, missing=missing_by_year)
ggplot(df_missing_by_year, aes(x=year, y=missing)) + 
  geom_point() +
  theme(text = element_text(size = tsize)) +
  ggtitle("Number of missing values in each year")

```


From the figure, we can see that the missing count is segmented roughly in decades. 1960s have the highest missing rate and the recent two years also miss many data. 1970s and 1980 have moderate high missing rate and 2000-2018 have the lowest missing rate.  It is interesting to see that the number of missing data drop sharply at around 1990. After looking into the data, the reason lies in the disassembly of the Soviet Union, which adds several countries that doesn't exist in the data set and lead to a fluctuation in records.However, even the lowest missing count is more 20 thousand, almost 1/3 of 60,543 records. Therefore, we concentrate on the time period 1971-2018 and we have to continue to check out the missing condition in each indicator.


```{r ch04-2}
df_count_sel_miss_y <- df_countryData_selected %>% 
  select(!c("1960":"1970", "2019":"2020"))
```

## Missing values by indicator
We then investigate the difference of missing values by indicators. We draw a plot that shows the distribution of average missing amount in 1970-2018 of each indicator.

```{r ch04-3}
df_count_sel_miss_y_isna <- cbind(df_count_sel_miss_y %>% select(c(1,2,3,4)), 
                                df_count_sel_miss_y %>% select(!c(1,2,3,4)) %>% is.na())

missing_indicator <- df_count_sel_miss_y_isna %>% 
  select(!c(1,2,4)) %>% 
  group_by(`Indicator Name`) %>% 
  summarise(across("1971":"2018", sum)) %>%  
  ungroup()

missing_indicator <- missing_indicator %>% 
  mutate(mean=rowMeans(missing_indicator %>% select(!1)))

ggplot(missing_indicator, aes(x=mean)) +
  geom_histogram(binwidth=12.5, fill="lightgray", color="black") +
  theme(text = element_text(size = tsize)) +
  ggtitle("Average missing amount by indicator")

```

We can see that there are many "high quality" indicators that have less than 50 missing values on average. However, there are also many indicators have more than 100 missing values. We drop the indicators that miss more than 150 values on average, since the remaining ones can not support a comprehensive investigation on them. 198 out of 279 selected indicators are left.

```{r ch04-4}
missing_indicator_l150 <- missing_indicator %>% 
  filter(mean < 150)

df_count_sel_miss_yi <- df_count_sel_miss_y %>% 
  filter(`Indicator Name` %in% missing_indicator_l150$`Indicator Name`)
```


## Missing values by country
At last, we examine the missing condition in each country.We decide to facet our figure with income group, since the difficulty of data collection differs for countries with different income level. 

```{r ch04-5}

df_only_countries <- read_csv("./WDI_csv/WDI_only_countries.csv")
df_count_sel_miss_yi_isna <- cbind(df_count_sel_miss_yi %>% select(c(1,2,3,4)), 
                                   df_count_sel_miss_yi %>% select(!c(1,2,3,4)) %>% is.na())

missing_country <- df_count_sel_miss_yi_isna %>% 
  select(!c(2,3,4)) %>% 
  group_by(`Country Name`) %>% 
  summarise(across("1971":"2018", sum)) %>%  
  ungroup()

missing_country <- missing_country %>% 
  mutate(mean=rowMeans(missing_country %>% select(!1)))

df_only_countries$`Income Group`[is.na(df_only_countries$`Income Group`)] <- "Low income"
missing_country <- merge(df_only_countries %>% 
                           select(c("Table Name", "Income Group")),  
                         missing_country, 
                         by.x="Table Name", by.y="Country Name")

missing_country$`Income Group` <- missing_country$`Income Group` %>% 
  factor(levels=c("Low income", "Lower middle income", "Upper middle income", "High income"))

ggplot(missing_country, aes(x=mean)) +
  geom_histogram(binwidth=12.5, fill="lightgray", color="black") +
  facet_wrap(~`Income Group`) +
  theme(text = element_text(size = tsize)) +
  theme(panel.spacing.x = unit(2, "lines")) +
  ggtitle("Average missing value by country, faceted by income") 

```

It was a little bit surprising that most of the countries that have high level of missing data are high income countries, which are supposed to have more robust data collecting systems. Luckily, the total number of high income country is large and it doesn't have much influence to remove some of them. Therefore, we only retain the countries that have less than 150 average missing values. After that, 197 out of 217 countries remains.

```{r ch04-6}

missing_country_l150 <- missing_country %>% 
  filter(mean < 150)

df_count_sel_miss_yic <- df_count_sel_miss_yi %>% 
  filter(`Country Name` %in% missing_country_l150$`Table Name`)

df_series_selected <- read_csv("./WDI_csv/WDI_series_selected.csv")
write_csv(df_count_sel_miss_yic, "./WDI_csv/WDI_Data_missing.csv")
colnames(missing_country_l150)[1] <- "Country Name"
write_csv(missing_country_l150 %>% select(c(1,2), "mean"), "./WDI_csv/selected_countries.csv")

df_series_selected$Topic <- df_series_selected$Topic %>% factor()
write_csv(df_series_selected %>% select(1,2,3) %>% filter(`Indicator Name` %in% missing_indicator_l150$`Indicator Name`) %>% 
            merge(missing_indicator_l150 %>% select(1, "mean"), by="Indicator Name"), "./WDI_csv/selected_indicators.csv") 
```
