==> 05-results.Rmd <==
# Results

```{r ch05-0}
library(tidyverse)
library(gridExtra)
df_data <- read_csv("./WDI_csv/WDIData.csv") %>% select(!last_col())
df_data_missing <- read_csv("./WDI_csv/WDI_Data_missing.csv")
df_selected_country <- read_csv("./WDI_csv/selected_countries.csv")
df_selected_indicator <- read_csv("./WDI_csv/selected_indicators.csv")
df_selected_country$`Income Group` <- df_selected_country$`Income Group` %>% 
  factor(levels=c("Low income", "Lower middle income", "Upper middle income", "High income"))
tsize <- 15
```

## Does the increase of education attainment and better overall economic condition increase overall well-being of a country?

Originally, we were going to investigate whether better education attainment and economic condition can decrease income inequality. However, indicators for income distribution, such as `Income share held by lowest 20%`, `Income share held by highest 20%`, have a lot of missing values. It seems that only about 30 countries are collecting data for this indicator, making it impossible for us to infer on the general trend worldwide. Therefore, we turn to `Life expectancy at birth, total` as a substitute since it can reflect the overall well-being of a country.

We would first like to investigate the relationship between education attainment and life-expectancy. However, most of the indicators for education attiment have a high percentage of missing values for most countries, again making it difficult for us to infer a general trend. The indicator `Primary completion rate, total (% of relevant age group)` seems to be the most viable one for analysis of general trend and we further investigate how its missing values are distributed here.

```{r ch05-1-1, fig.width=10}
# name the chunks to avoid conflict
df_edu_temp <- merge(df_selected_country %>% select(c(1,2)), 
                     cbind(df_selected_country %>% select("Country Name"),
                       df_data_missing %>% filter(`Indicator Code` == "SE.PRM.CMPT.ZS") %>% 
                       select(!c(1,2,3,4)) %>% is.na()),
                     by="Country Name")
edu_missing <-  df_edu_temp %>% 
  group_by(`Income Group`) %>% 
  summarise(across(`1971`:`2018`, mean)) %>% 
  ungroup() %>% 
  pivot_longer(!c("Income Group"), names_to="Year", values_to="Missing Rate")
edu_missing$`Income Group` <- edu_missing$`Income Group` %>% 
  factor(levels=c("Low income", "Lower middle income", "Upper middle income", "High income"))
edu_missing$Year <- as.integer(edu_missing$Year)
ggplot(edu_missing, aes(x=Year, y=`Missing Rate`, color=fct_rev(`Income Group`))) +
  geom_line() +
  theme(text = element_text(size = tsize)) +
  scale_color_discrete(name="Income Group") +
  ggtitle("Missing Rate of Primary Completion Rate Records")
```

  We can observe that the number of missing values dropped sharply for all conturies after 1998, which might be attributed to the attention education attainment received in recent years. As a result, we decide to limit our analysis for question to the time period betwen 1999 and 2017. In addition, the rate of values are similar for all four income groups, so dropping countries without any data for this indicator will not lead to a biased distribution between income groups.



To investigate on the relationship between education and life expectancy, we first draw a plot between them at year 2010.

```{r ch05-1-2, fig.width=10}
df_life_exp <- merge(df_selected_country %>% select(c(1,2)), 
                     cbind(df_selected_country %>% select("Country Name"),
                           df_data %>% 
                             filter(`Indicator Code` == "SP.DYN.LE00.IN") %>% 
                             filter(`Country Name` %in% df_selected_country$`Country Name`) %>% 
                             select(df_data_missing %>% colnames()) %>% 
                             select(!c(1,2,3,4))),
                     by="Country Name")
df_edu <- merge(df_selected_country %>% select(c(1,2)), 
                     cbind(df_selected_country %>% select("Country Name"),
                           df_data_missing %>% filter(`Indicator Code` == "SE.PRM.CMPT.ZS") %>% 
                             select(!c(1,2,3,4))),
                     by="Country Name")
life_exp_2010 <- df_life_exp %>% select(c(`Country Name`, `2010`))
colnames(life_exp_2010)[2] <- "Life Expectancy"
edu_2010 <- df_edu %>% select(c(`Country Name`, `2010`))
colnames(edu_2010)[2] <- "Primary Completion Rate"
df_life_edu_2010 <- merge(life_exp_2010, edu_2010, by="Country Name")
ggplot(df_life_edu_2010, aes(x=`Primary Completion Rate`, y=`Life Expectancy`)) +
  geom_point(na.rm=TRUE) +
  geom_smooth(na.rm=TRUE, method=lm, se=FALSE) +
  theme(text = element_text(size = tsize)) +
  ylab("Life Expectancy, years") +
  xlab("Primary Completion Rate, %") +
  ggtitle("Primary Completion Rate vs Life Expecctancy in 2010")
  
```

From the plot, we can see a strong positive correlation between primary completion rate and the life expectancy. Here we need to note that the primary completion rate can be more than 100 since it is the fraction of students enrolled in the last year divided by the students enrolled in the first year. As a result, reapeating students and transfer students will lead to a higher completion rate than expected. Although we can not conclude that higher primary complete rate directly leads to higher life expectancy, we can conclude that they are highly correlated and are both indicators of the overall well-being of a country.

After that, we would like to see how primary complete rate and life expectancy have changed over time.

```{r ch05-1-3, fig.width=10}
groups=c("Low income", "Lower middle income", "Upper middle income", "High income")
df_life_exp_group <- df_data %>% 
                             filter(`Indicator Code` == "SP.DYN.LE00.IN") %>% 
                             filter(`Country Name` %in% groups) %>% 
                             select(df_data_missing %>% colnames()) %>% 
                             select(!c(2,3,4))
df_life_exp_group$`Country Name` <- df_life_exp_group$`Country Name` %>% 
  factor(levels=c("Low income", "Lower middle income", "Upper middle income", "High income"))
df_life_exp_group <- df_life_exp_group %>% 
  select(!c("1971":"1998","2018")) %>% 
  pivot_longer(-1, values_to="life", names_to="year")
df_edu_group <- df_data %>% 
  filter(`Indicator Code` == "SE.PRM.CMPT.ZS") %>% 
  filter(`Country Name` %in% groups) %>% 
  select(df_data_missing %>% colnames()) %>% 
  select(!c(2,3,4))
df_edu_group$`Country Name` <- df_edu_group$`Country Name` %>% 
  factor(levels=c("Low income", "Lower middle income", "Upper middle income", "High income"))
df_edu_group <- df_edu_group %>% 
  select(!c("1971":"1998","2018"))%>% 
  pivot_longer(-1, values_to="edu", names_to="year")
df_life_edu_group <- merge(df_life_exp_group, df_edu_group, by=c("Country Name", "year"))
df_life_edu_group$year <- as.integer(df_life_edu_group$year)
ggplot(df_life_edu_group) +
  theme(text = element_text(size = tsize)) +
  geom_line(mapping=aes(x=year, y=edu, color=fct_rev(`Country Name`))) +
  geom_point(mapping=aes(x=year, y=life, color=fct_rev(`Country Name`))) +
  xlab("Year") +
  ylab("Primary Completion Rate, line, %\n Life Expectancy, dot, year") +
  scale_color_discrete(name="Income Group") +
  ggtitle("Change of primary completion rate (line, %)\nand life expectancy (dot, years)")
```

From the plot we can see that higher income groups have much higher primary completion rate and life expectancy. High income and upper middle income countries maintains near 100% primary completion rate our window of study, while primary complete rate in lower middle and low income countries grow rapidly thanks to the increasing focus of fundamental education in these areas.


```{r pivot dataset}
# pivot the data from wide to long
country_income_group <- df_selected_country %>%
  select(`Country Name`,`Income Group`)
df_panel <- df_data_missing %>% 
  # first wrangle column of years into rows
  pivot_longer(cols = "1971":"2018",names_to = "Year")%>% 
  # then pivot rows of indicator into columns
  pivot_wider(id_cols = c(`Country Name`,Year),names_from = `Indicator Code`,values_from = value)%>%
  left_join(country_income_group,by = "Country Name")%>%
  mutate(Year = as.numeric((Year)))%>%
  select(`Country Name`,Year,`Income Group`,everything())

df_indicators <- df_data_missing %>%
  distinct(`Indicator Name`,`Indicator Code`)
```

##  Is there a co-relation between CO2 emissions and financial indicators, quality of life, etc?

We first examine the emission of greenhouse gases per capita for each income group. We choose the year 2000 to draw a boxplot showing the overall distribution.
```{r ch05-2-1, fig.height=5, fig.width=10}
df_2010_co2 <- merge(df_selected_country %>% select(c(1,2)), 
                     df_data_missing %>% filter(`Indicator Code` == "EN.ATM.CO2E.PC") %>% 
                       select(c("Country Name", "2010")),
                     by="Country Name")

colnames(df_2010_co2)[3] <- "co2 per capita"
ggplot(df_2010_co2, aes(x=`co2 per capita`, y=`Income Group`)) + 
  geom_boxplot(na.rm=TRUE) +
  theme(text = element_text(size = tsize)) +
  xlab("CO2 emission per capita in 2010 (tons)") + 
  ggtitle("Boxplot of CO2 emission per capita")

```

From the plot, we can see the evidence that countries with higher income emit more co2 per person. Most of the low-income countries emit almost negligible co2 compared to high-income countries. However, this might be due to the fact that the production of high-value product requires more energy and low-income countries tend to have bigger population. Therefore, we also plot the relationship between greenhouse emission per 2015 US$ and income groups.

```{r ch05-2-2, fig.height=5, fig.width=10}
df_2010_co2 <- merge(df_2010_co2, 
                     df_data_missing %>% filter(`Indicator Code` == "EN.ATM.CO2E.KD.GD") %>% 
                       select(c("Country Name", "2010")),
                     by="Country Name")

colnames(df_2010_co2)[4] <- "co2 per 2015 US$"
ggplot(df_2010_co2, aes(x=`co2 per 2015 US$`, y=`Income Group`)) + 
  geom_boxplot(na.rm=TRUE) +
  theme(text = element_text(size = tsize)) +
  xlab("CO2 emission per 2015 US$ in 2010 (kg)") +
  # scale_y_discrete(position = "right") +
  ggtitle("Boxplot of CO2 emission per 2015 US$")
```

This plot gives us some different insight than the former. The plot shows that high and low income countries emit relatively fewer co2 to generate a unit of GDP compared to upper middle and lower middle income countries. The reason may comes from the fact that high-income countries have a wholesome system of service sector that can make a lot of profit with relatively low pollution while the low-income countries focus on fundamental livings which also emit few co2. In contrast, middle-income countries may focus on industrial section, which will emit high amount of co2 for each unit of GDP.

At last, we would like see the annually varying situation of the two above-mentioned metrics.

```{r ch05-2-3, fig.width=10}
df_co2 <- df_data_missing %>% 
  filter(`Indicator Code` == "EN.ATM.CO2E.KT") %>% 
  select(!c("Country Code", "Indicator Name", "Indicator Code")) %>% 
  column_to_rownames("Country Name")
df_population <- df_data_missing %>% 
  filter(`Indicator Code` == "SP.POP.TOTL") %>% 
  select(!c("Country Code", "Indicator Name", "Indicator Code")) %>% 
  column_to_rownames("Country Name")
df_gdp <- df_data_missing %>% 
  filter(`Indicator Code` == "NY.GDP.MKTP.KD") %>% 
  select(!c("Country Code", "Indicator Name", "Indicator Code")) %>% 
  column_to_rownames("Country Name")

df_co2_temp <- df_co2 * !is.na(df_population)
df_population_temp <- df_population * !is.na(df_co2)
df_co2_temp <- merge(df_selected_country %>% select(c("Country Name", "Income Group")), 
                     df_co2_temp %>% rownames_to_column("Country Name"),
                     by="Country Name")
df_population_temp <- merge(df_selected_country %>% select(c("Country Name", "Income Group")), 
                            df_population_temp %>% rownames_to_column("Country Name"),
                            by="Country Name")

df_co2_group <- df_co2_temp %>% 
  select(!1) %>% 
  group_by(`Income Group`) %>% 
  summarise(across("1971":"2018", sum, na.rm=TRUE)) %>%  
  ungroup()

df_population_group <- df_population_temp %>% 
  select(!1) %>% 
  group_by(`Income Group`) %>% 
  summarise(across("1971":"2018", sum, na.rm=TRUE)) %>%  
  ungroup()

df_co2_pc_group <- (df_co2_group %>% column_to_rownames("Income Group") / 
  df_population_group %>% column_to_rownames("Income Group") * 1000) %>% 
  rownames_to_column("Income Group") %>%
  pivot_longer(cols=-1, names_to="year")

df_co2_pc_group$`Income Group` <- df_co2_pc_group$`Income Group` %>% 
  factor(levels=c("Low income", "Lower middle income", "Upper middle income", "High income"))

df_co2_pc_group$year <- as.integer(df_co2_pc_group$year)
ggplot(df_co2_pc_group, aes(x=year, y=value, color=fct_rev(`Income Group`))) +
  geom_line() +
  theme(text = element_text(size = tsize)) +
  ylab("CO2 emission per capita (ton)") +
  scale_color_discrete(name="Income Group") +
  ggtitle("Change of CO2 emission per capita")
```


We can see that the trend is close to what is shown in the boxplot of CO2 emission per capita, which increases with the increase of income group. We can also observe that the emission per capita in high-income countries drops in the recent years, which may credits to the development of service section and the arise of environment-protect activities. On the other hand, the emission in upper-middle-income countries bursts after 2000, because of the rapid industrialization in these countries. Also, we can observe that the emission of lower-middle-income countries and low-income countries have a cross at about 1990, while the record of upper middle income countries drop sharply at the same time, because the additional data brought by the disassembly of Soviet Union.

Similarly, we want to discover how CO2 emission per 2015 US$ has changed over time.

```{r ch05-2-4, fig.width=10}
df_co2_temp <- df_co2 * !is.na(df_gdp)
df_gdp_temp <- df_gdp * !is.na(df_co2)
df_co2_temp <- merge(df_selected_country %>% select(c("Country Name", "Income Group")), 
                     df_co2_temp %>% rownames_to_column("Country Name"),
                     by="Country Name")
df_gdp_temp <- merge(df_selected_country %>% select(c("Country Name", "Income Group")), 
                            df_gdp_temp %>% rownames_to_column("Country Name"),
                            by="Country Name")

df_co2_group <- df_co2_temp %>% 
  select(!1) %>% 
  group_by(`Income Group`) %>% 
  summarise(across("1971":"2018", sum, na.rm=TRUE)) %>%  
  ungroup()

df_gdp_group <- df_gdp_temp %>% 
  select(!1) %>% 
  group_by(`Income Group`) %>% 
  summarise(across("1971":"2018", sum, na.rm=TRUE)) %>%  
  ungroup()

df_co2_pdollar_group <- (df_co2_group %>% column_to_rownames("Income Group") / 
                    df_gdp_group %>% column_to_rownames("Income Group") * 10^6) %>% 
  rownames_to_column("Income Group") %>%
  pivot_longer(cols=-1, names_to="year")

df_co2_pdollar_group$`Income Group` <- df_co2_pdollar_group$`Income Group` %>% 
  factor(levels=c("Low income", "Lower middle income", "Upper middle income", "High income"))

df_co2_pdollar_group$year <- as.integer(df_co2_pdollar_group$year)
ggplot(df_co2_pdollar_group, aes(x=year, y=value, color=fct_rev(`Income Group`))) +
  geom_line() +
  theme(text = element_text(size = tsize)) +
  ylab("CO2 emission per 2015 US$ (kg)") +
  scale_color_discrete(name="Income Group") +
  ggtitle("Change of CO2 emission per 2015 US$")
```

From the plot, we can reaffirm our previous observation that middle income countries emits more CO2 to produce a unit of GDP while high-income and low-income emits less. We also see a sharp spike at about 1990 in upper-middle and lower-middle-income countries, which is also due to the dissolution of the Soviet Union. In addition, except for the low-income countries, the world shows a steady drop of CO2 emission per dollar after 1990, owing to the fact that the development of technology can reduce the CO2 emission per product.

## Does richer countries use more renewable sources for electricity production and does this indeed reduce the emission of CO2?

We first investigate whether countries with higher income generate more electricity using renewable sources. Here, we define both nuclear sources and all other sources, which includes hydroelectric, solar, wind etc. as renewable sources for electricity production. The indicator we use are the percentage of electricity produced from nuclear sources and percentage of electricity produced from other sources. We then sum them to get the total percentage of electricity produced from renewable sources. We then aggregate them to yearly income group level by taking the mean. Moreover, we limit this analysis to the period between 1990 and 2016, since the indicator contains mostly missing values outside of this date range. 

We can draw several interesting conclusion from the plot below. Firstly,low income countries have the highest percentage of electricity produced from renewable sources. However, these indicators contains high counts of missing values for low income countries, so this observation might not be robust. Secondly, percentage of electricity produced from renewable sources remains mostly constant for upper middle income countries and exhibit a downward trend for lower middle income countries, while this percentage has been rapidly increasing for high income and low income countries since around 2006. This could be due to increasing awareness of the importance of renewable energy since this period. Third, only rich countries can afford to use nuclear energy as a source of electricity, while this source electricity remains below 5% of total electricity production during this time period for all other countries.

```{r ch05-3-1, fig.width=10}

df_renewable_energy <- df_panel %>%
  group_by(`Income Group`,Year)%>%
  summarize(`Other Sources` = mean(EG.ELC.RNEW.ZS,na.rm = TRUE),`Nuclear Sources` = mean(EG.ELC.NUCL.ZS,na.rm = TRUE))%>%
  mutate(`Other Sources` = replace_na(`Other Sources`,0),`Nuclear Sources` = replace_na(`Nuclear Sources`,0))%>%
  mutate(`All Renewable Sources` = `Other Sources` + `Nuclear Sources`)%>%
  pivot_longer(cols = `Other Sources`:`All Renewable Sources`,names_to = "Renewable Energy Type")%>%
  mutate(`Income Group` = fct_rev(`Income Group`))%>%
  ungroup()%>%
  filter(Year >1990 & Year < 2016)%>%
  ungroup()
 
# electricity production from renewable sources, break down by country and type  
df_renewable_energy%>%
  ggplot(aes(x = Year,y = value,color = `Renewable Energy Type`)) + 
    geom_point(position=position_jitter(w=0.2),alpha = 0.8)+
  geom_line(position=position_jitter(w=0.2),alpha = 0.8)+
  facet_grid(~`Income Group`)+
  scale_x_continuous(breaks = round(seq(1971, 2018, by = 10))) +
  theme(legend.position = "bottom",text = element_text(size = tsize))+
  labs(x = " ",y = "Percentage")+
  ggtitle("Electricity Production from Renewable Sources",subtitle = "% of total production")
    
```

To investigate the relationship between use of renewable electricity source and CO2 emission, we first investigate the co-movement between these indicators. Since CO2 emission indicators are respectively in kg and tons, we scale CO2 emission indicators using min max scaling w.r.t each income group so that all indicators in the plot are on similar scales to help with our visualization. For high income countries and low income countries, increase in use of renewable electricity corresponds to a decrease in both indicators for CO2 emission. However, CO2 emissions per capita increase while CO2 per dollar decreased for upper and lower middle income countries, which suggests that other factors, such as productivity increase and/or consumption increase also affects CO2 emissions.

```{r ch05-3-2, fig.width=10}
 
# does renewable energy corresponds to decrease in emission?

# EN.ATM.CO2E.KD.GD	
# EN.ATM.CO2E.PC
# 
# EN.ATM.NOXE.EG.ZS
# 	EN.CO2.ETOT.ZS

df_emission <- df_panel %>%
  group_by(`Income Group`,Year)%>%
  summarize(`CO2 per 2015 USD` = mean(EN.ATM.CO2E.KD.GD,na.rm = TRUE),
            `CO2 per Capita` = mean(EN.ATM.CO2E.PC,na.rm = TRUE))%>%
  mutate(`CO2 per 2015 USD` = replace_na(`CO2 per 2015 USD`,0),
         `CO2 per Capita` = replace_na(`CO2 per Capita`,0),
         `Income Group` = fct_rev(`Income Group`))

df_3_2<-df_renewable_energy %>%
  filter(`Renewable Energy Type` == "All Renewable Sources")%>%
  mutate(all_renewable = value)%>%
  select(`Income Group`,Year,all_renewable)


emission_and_renewable <- df_emission %>%
  left_join(df_3_2,by = c("Income Group","Year"))%>%
  filter(Year >1990 & Year < 2016)%>%
  # convert to 0 - 1
  mutate(all_renewable = all_renewable/100)%>%
  # min max scaling for emission
  mutate_at(vars(`CO2 per 2015 USD`:`CO2 per Capita`),~scales::rescale(.x,to = c(0,1)))%>%
  ungroup()

emission_and_renewable%>%
  pivot_longer(cols = `CO2 per 2015 USD`:all_renewable,names_to = "series name")%>%
  ggplot(aes(x = Year,y = value,color = `series name`))+
  geom_point()+
  geom_smooth(method = "loess",se=FALSE,span = 0.4,lwd = 0.5)+
  facet_grid(~`Income Group`)+
  scale_x_continuous(name= " ",breaks = round(seq(1990, 2016, by = 10))) +
  theme(legend.position = "bottom",text = element_text(size = tsize))+
  labs(y = "Standardized Value")+
  scale_color_discrete(name = " ", labels = c("Renewable Electricity %", "CO2 per 2015 USD", "CO2 per Capita"))+
  ggtitle("Comparison of renewable energy source and CO2 emission")

```

We investigate this issue further by visualizing the data at the more granular country level. We created a scatter plot and fit a least square line between renewable electricity percentage and both CO2 emission per capita and CO2 emission per dollar for all income groups. We choose to log CO2 emissions indicators to again account for scale differences. For income groups, there exists a strong negative relationship between renewable electricity percentage and CO2 emission, although the large spread around the least square line indicates that there are other potential explanatory variables for CO2 emission. 

```{r ch05-3-2-1, fig.width=10}


g1<-df_panel %>% 
  select(Year,`Country Name`,`Income Group`,EG.ELC.RNEW.ZS,EG.ELC.NUCL.ZS,EN.ATM.CO2E.KD.GD,EN.ATM.CO2E.PC)%>%
  filter(Year >1990 & Year < 2016)%>%
  drop_na()%>%
  rename(other_percent = EG.ELC.RNEW.ZS,nuclear_percent = EG.ELC.NUCL.ZS,co2_per_usd = EN.ATM.CO2E.KD.GD,co2_capita = EN.ATM.CO2E.PC)%>%
  mutate(renewable_percentage = (other_percent + nuclear_percent)/100)%>%
  ggplot(aes(x = renewable_percentage,y = log(co2_per_usd)))+
  geom_point()+
  geom_smooth(method = 'lm')+
  facet_grid(~`Income Group`)+
  labs(x= "Electricity Production from Renewable Sources,(% of total)",y = "CO2 per 2015 USD,(log(kg))")+
  ggtitle("Correlation Between Renewable Electricity Percentage and CO2 per 2015 USD")

g2<-df_panel %>% 
  select(Year,`Country Name`,`Income Group`,EG.ELC.RNEW.ZS,EG.ELC.NUCL.ZS,EN.ATM.CO2E.KD.GD,EN.ATM.CO2E.PC)%>%
  filter(Year >1990 & Year < 2016)%>%
  drop_na()%>%
  rename(other_percent = EG.ELC.RNEW.ZS,nuclear_percent = EG.ELC.NUCL.ZS,co2_per_usd = EN.ATM.CO2E.KD.GD,co2_capita = EN.ATM.CO2E.PC)%>%
  mutate(renewable_percentage = (other_percent + nuclear_percent)/100)%>%
  ggplot(aes(x = renewable_percentage,y = log(co2_capita)))+
  geom_point()+
  geom_smooth(method = 'lm')+
  facet_grid(~`Income Group`)+
  labs(x= "Electricity Production from Renewable Sources,(% of total)",y = "CO2 per capita,(log(kg))")+
  ggtitle("Correlation Between Renewable Electricity Percentage and CO2 per capita")

gridExtra::grid.arrange(g1,g2)
  
```
