==> 06-interactive.Rmd <==
# Interactive component

As is shown in the answering the question, for example, in the process of drawing the relationship between primary completion rate and life expectancy, we only plot a figure concentrating on the data of year 2010, since we can not display the data of every year in one figure. With the help of interactive plots, we now have the opportunity to show the data from all years by adding a slide bar to control the year to display. In addition, the radius of the circle represents the population of the country and color shows the income group it belongs to. From the interactive plot, we can see a positive correlation between primary school completion rate and life expectancy in every year.

<meta charset="utf-8">
        
<!-- Load d3.js -->
<script src="https://d3js.org/d3.v6.js"></script>

<input type="range" name="yearSlider" id=yearSlider min="1971" max="2018" value="2018" style="width:50%; background:#d3d3d3" oninput="this.nextElementSibling.value = this.value">
<output>2018</output>

<p id="country_name">Hover over a circle to identify the country...</p>

<!-- Create a div where the graph will take place -->
<div id="my_dataviz"></div>

<script>

const radius = 4;

const color_lookup = {'Low income':'#A20000',
					  'Lower middle income': '#A25300',
					  'Upper middle income': '#14A200',
					  'High income': '#0014A2'};

// set the dimensions and margins of the graph
const margin = {top: 30, right: 60, bottom: 60, left: 50},
      width = 840 - margin.left - margin.right,
      height = 680 - margin.top - margin.bottom;
    
// append the svg object to the body of the page
const svg = d3.select("#my_dataviz")
    .append("svg")
    .attr("width", width + margin.left + margin.right)
    .attr("height", height + margin.top + margin.bottom)
    .append("g")
    .attr("transform", `translate(${margin.left}, ${margin.top})`);

// add X axis
const x = d3.scaleLinear()
 			.domain([20, 100])
			.range([0, width]);

svg.append("g")
   .attr("transform", `translate(0, ${height})`)
   .call(d3.axisBottom(x));

// add Y axis
const y = d3.scaleLinear()
            .domain([0, 140])
            .range([height, 0]);

svg.append("g")
   .call(d3.axisLeft(y));

// add the X gridlines
svg.append("g")
   .attr("class", "grid")
   .attr("transform", "translate(0," + height + ")")
   .call(d3.axisBottom(x)
   	       .ticks(10)
   		   .tickSize(-height)
   		   .tickFormat(""))
   .style("stroke", "lightgrey")
   .style("stroke-opacity", 0.2)
   .style("stroke-width", 1);

// add the Y gridlines
svg.append("g")			
   .attr("class", "grid")
   .call(d3.axisLeft(y)
   		   .ticks(10)
           .tickSize(-width)
           .tickFormat(""))
   .style("stroke", "lightgrey")
   .style("stroke-opacity", 0.2)
   .style("stroke-width", 1);

// title
svg.append('text')
    .attr('x', width/2)
    .attr('y', -10)
    .attr('text-anchor', 'middle')
    .style('font-family', 'Helvetica')
    .style('font-size', 20)
    .text('Primary Completion Rate vs Life Expectancy');
    
// X label
svg.append('text')
   .attr('x', width/2)
   .attr('y', height + 40)
   .attr('text-anchor', 'middle')
   .style('font-family', 'Helvetica')
   .style('font-size', 16)
   .text('Primary Completion Rate (percentage)');

// Y label
svg.append('text')
   .attr('x', 250)
   .attr('y', -90)
   .attr('text-anchor', 'middle')
   .attr('transform', 'translate(60,' + height + ')rotate(-90)')
   .style('font-family', 'Helvetica')
   .style('font-size', 16)
   .text('Life Expectancy (years)');

// Add Legend
var legend = svg.selectAll(".legend")
 				.data(["High income", "Upper middle income", "Lower middle income", "Low income"])
 				.enter()
 				.append("g")
 				.attr("class", "legend")
 				.attr("transform", function(d, i) { return "translate(0," + i * 20 + ")"; });

legend.append("rect")
	  .attr("x", width - 18)
	  .attr("y", height - 95)
 	  .attr("width", 18)
 	  .attr("height", 18)
 	  .style("fill", function(d) { return color_lookup[d]} );

legend.append("text")
 	  .attr("x", width - 24)
 	  .attr("y", height - 88)
 	  .attr("dy", ".35em")
 	  .style("text-anchor", "end")
 	  .text(function(d) { return d; });

d3.csv("https://raw.githubusercontent.com/Zhao-Wenbo/EDAV_final_project/main/interactive/data.csv").then(
	function(csv) {

	csv = csv.filter(function(row) {return row['life_exp'] !== 'NA'});
	csv = csv.filter(function(row) {return row['primary_completion'] !== 'NA'});
	csv = csv.filter(function(row) {return row['population'] !== 'NA'});

	filtered_data = csv.filter(function(row) {return row['year'] === "2018"});

    // Add dots
    svg.append('g')
	   .selectAll("circle")
	   .data(filtered_data)
	   .join("circle")
	   .attr("cx", function (d) { return x(d.life_exp); })
	   .attr("cy", function (d) { return y(d.primary_completion); })
	   .attr("country", function (d) {return d.country; })
	   .attr("region", function (d) {return d.region; })
	   .attr("r", function (d) { return determineDotSize(d.population) })
	   .style("fill", function (d) { return color_lookup[d.income]; })
	   .on("mouseover", function(d, i) {
		  d3.select(this).attr("stroke", "black").attr("stroke-width", 5);
		  document.getElementById("country_name").innerHTML = this.getAttribute("country") + ' (' + this.getAttribute("region") + ')';
	   })
       .on("mouseout", function(d, i) {
       		d3.select(this).attr("stroke", "transparent").attr("stroke-width", 0)
       		document.getElementById("country_name").innerHTML = "Hover over a circle to identify the country...";
       		d3.select("#temp").remove();
       });

    // Add Regression Line
    var xData = filtered_data.map(function(d) { return d.life_exp; }).map(Number);
    var yData = filtered_data.map(function(d) { return d.primary_completion; }).map(Number);

	regression = leastSquaresEquation(xData, yData);

	var line = d3.line()
    		   .x(function(d) { return x(d.life_exp); })
               .y(function(d) { return y(regression(d.life_exp)); });

    svg.append("path")
      .datum(filtered_data)
      .attr("class", "line")
      .attr("d", line)
      .attr("id", "regress_line")
      .style("fill", "none")
      .style("stroke", "black")
      .style("stroke-width", "4");

  	// Listen to the slider
	d3.select("#yearSlider").on("change", function(d) {
		year = parseInt(this.value)
		filtered_data = csv.filter(function(row) { return parseInt(row['year']) === year; })

		// update the chart
	    svg.selectAll("circle")
	       .data(filtered_data)
		   .attr("cx", function (d) { return x(d.life_exp); })
		   .attr("cy", function (d) { return y(d.primary_completion); })

		var xData = filtered_data.map(function(d) { return d.life_exp; }).map(Number);
    	var yData = filtered_data.map(function(d) { return d.primary_completion; }).map(Number);

	    // Add regression line
		regression = leastSquaresEquation(xData, yData);

		var line = d3.line()
    		   	 	 .x(function(d) { return x(d.life_exp); })
               		 .y(function(d) { return y(regression(d.life_exp)); });

    	svg.select("#regress_line")
       	   .datum(filtered_data)
           .attr("class", "line")
       	   .attr("d", line)
	});

	function determineDotSize(population) {
		if (population > 100000000) {
			return (radius * 3);
		} else if (population > 20000000) {
			return (radius * 2);
		} else if (population < 5000000) {
			return (radius & 0.5);
		} else { return radius; }
	}

	function leastSquaresEquation(xData, yData) {
    	var ReduceAddition = function(prev, cur) { return prev + cur; };
    
    	// finding the mean of Xaxis and Yaxis data
    	var xBar = xData.reduce(ReduceAddition) * 1.0 / xData.length;
    	var yBar = yData.reduce(ReduceAddition) * 1.0 / yData.length;

    	console.log(xBar)
    	console.log(yBar)

    	var SquareXX = xData.map(function(d) { return Math.pow(d - xBar, 2) }).reduce(ReduceAddition)
    	var ssYY = yData.map(function(d) { return Math.pow(d - yBar, 2); }).reduce(ReduceAddition);

    	var MeanDiffXY = xData.map(function(d, i) { return (d - xBar) * (yData[i] - yBar); }).reduce(ReduceAddition);
      	var slope = MeanDiffXY / SquareXX;
      	var intercept = yBar - (xBar * slope);

    	// returning regression function
    	return function(x) { return ((x * slope) + intercept); }
  }

})
</script>
