==> 07-conclusion.Rmd <==
# Conclusion

From the first question we answered, we can see that we change the original topic of income inequality to life expectancy, since the former have too much missing data. Therefore we should have much more detailed investigation on the completeness of the data before start to work. In addition, we found that `read_csv` from `dplyr` is much more efficient than `read.csv` from `base` package. The former takes almost no time read a file of several hundreds of megabits while the later takes several minutes and more importantly, the later will not be bothered with the data encoding. Hope that the data of income inequality can be completely collected so that we can study on the relation between education attainment and income inequality in the near future.

In the second question, we can see that both as a per unit metric, `CO2 emission per capita` and `CO2 emission per 2015 US$` give us completely different image. `CO2 emission per capita` shows that developed countries emit more greenhouse gases per residence while `CO2 emission per 2015 US$` shows that developed countries are more efficient in greenhouse emission on producing unit production. Therefore, this situation gives us a hint that many things can be viewed differently from various perspectives and we should not be stuck to a single opinion.

For the 3rd the question we asked, we found that while rich country does use more renewable energy to generate electricity, other countries also extensively use renewable energy to generate electricity. In fact, upper middle and lower middle income countries have been consistenly generating 30% ~ 40% percent of its electricity using renewable energy, while low income country generate close to 50% percent of its electricity in recent years. What does seperate rich countries from them is the use of nuclear energy to generate electricity. Rich countries produce around 10% of its electricity form nuclear energy, while other countries barely use any nuclear energy. This is likely due to the cost and technology barrier to using nuclear energy at scale. 

Additionally, we found there is likely strong negative correlation between the percentage of electricity generated by renewable energy and CO2 emission. In recent years, an increasing percentage of eletricity is being produced by renewable energy in both rich and low income countries, which seems to contribute to the continued drop in CO2 emissions in both countries. Moreover, the least-square lines we fit between percentage of electricity generated by renewable energy and CO2 emission has a obvious downward trend and small standard error, which also indicate a strong negative relationship between use of renewable energy and CO2.
